requirements.txt
ncclient>=0.5.2
lxml>=3.2.4
pyang
pyYAML
